+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: bda-silv <bda-silv@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/11 08:38:32 by bda-silv          #+#    #+#             */
-/*   Updated: 2022/06/01 08:15:48 by bda-silv         ###   ########.fr       */
+/*   Updated: 2022/06/09 14:14:41 by bda-silv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
