+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: bda-silv <bda-silv@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/01 10:50:24 by bda-silv          #+#    #+#             */
-/*   Updated: 2022/06/06 13:31:41 by bda-silv         ###   ########.fr       */
+/*   Updated: 2022/06/09 14:42:50 by bda-silv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,10 +34,10 @@
 
 #include "libft.h"
 
-char	*ft_strnstr(const char *hay, const char *needle, unsigned int n)
+char	*ft_strnstr(const char *hay, const char *needle, size_t n)
 {
-	unsigned int	i;
-	unsigned int	n_needle;
+	size_t	i;
+	size_t	n_needle;
 
 	n_needle = ft_strlen((char *)needle);
 	if (!*needle)
